Consolidate benchmark code: Liquid is the only path
Aligns the benchmark code with the data-generator change (staging tables
now CLUSTER BY). Removes the parallel partitioned/Liquid setup notebooks
and the dbt Jinja `use_liquid_clustering` gate — Liquid is the project
default everywhere.

Notebook consolidation:
  setup_liquid.py            -> setup.py            (overwrite)
  setup_dbt_liquid.py        -> setup_dbt.py        (overwrite)
  DLT/pipelines_setup_liquid.py  -> DLT/pipelines_setup.py
  DLT/dlt_ingest_bronze_liquid.py -> DLT/dlt_ingest_bronze.py
  DLT/dlt_historical_liquid.sql  -> DLT/dlt_historical.sql
  DLT/dlt_incremental_liquid.sql -> DLT/dlt_incremental.sql

All _liquid duplicates removed. The canonical filenames now contain
the Liquid behaviors. Modclust SDP variant (engineer-suggested cluster
keys) is intentionally preserved as a separate experiment.

dbt model changes:
  - Removed `{% if var('use_liquid_clustering', false) %}...{% else %}...
    {% endif %}` Jinja gate from all 7 bronze models and the 3 gold models
    that had one (factmarkethistory, factcashbalances, currentaccountbalances).
    Only the Liquid branch behavior remains.
  - factmarkethistory + factcashbalances: now always `merge` + composite
    `unique_key` (was `insert_overwrite` partitioned).
  - currentaccountbalances: always `insert_overwrite` no partition_by
    (CREATE OR REPLACE TABLE AS SELECT) — `latest_batch` boolean partition
    retired.
  - 7 bronze models: no layout config in dbt (setup-owns-layout pattern);
    `setup_dbt.py` pre-creates with CLUSTER BY + tblproperties.

Docs:
  - CLAUDE.md, src/dbt_augmented_incremental/README.md,
    src/incremental_batches/augmented_incremental/README.md updated to
    describe Liquid as the only path. Removed mentions of the deleted
    _liquid files and the use_liquid_clustering var.

Tests:
  - tests/test_workflow_builders.py comment update only. All 17 tests
    still pass.
