--- src/incremental_batches/augmented_incremental/DLT/dlt_ingest_bronze_liquid.py
@@ -1,117 +1,0 @@
-# Databricks notebook source
-import dlt
-
-# spark.conf.set("spark.sql.autoBroadcastJoinThreshold", 262144000)
-# spark.conf.set("spark.databricks.adaptive.autoBroadcastJoinThreshold", 262144000)
-
-# COMMAND ----------
-
-tbl_ls          = ['account', 'cashtransaction', 'customer', 'dailymarket', 'holdings', 'trade', 'watches']
-tpcdi_directory = spark.conf.get('tpcdi_directory') # Place this in the DLT JSON Config
-scale_factor    = spark.conf.get('scale_factor')    # Place this in the DLT JSON Config
-wh_db           = spark.conf.get('wh_db')           # Place this in the DLT JSON Config
-batches_dir     = f"{tpcdi_directory}augmented_incremental/_dailybatches/{wh_db}_{scale_factor}"
-
-schemas = {
-  "account": "cdc_flag STRING, cdc_dsn BIGINT, accountid BIGINT, brokerid BIGINT, customerid BIGINT, accountdesc STRING, taxstatus TINYINT, status STRING, update_dt DATE",
-  "cashtransaction": "cdc_flag STRING, cdc_dsn BIGINT, accountid BIGINT, ct_dts TIMESTAMP, ct_amt DOUBLE, ct_name STRING, event_dt DATE",
-  "customer": "cdc_flag STRING, cdc_dsn BIGINT, customerid BIGINT, taxid STRING, status STRING, lastname STRING, firstname STRING, middleinitial STRING, gender STRING, tier TINYINT, dob DATE, addressline1 STRING, addressline2 STRING, postalcode STRING, city STRING, stateprov STRING, country STRING, c_ctry_1 STRING, c_area_1 STRING, c_local_1 STRING, c_ext_1 STRING, c_ctry_2 STRING, c_area_2 STRING, c_local_2 STRING, c_ext_2 STRING, c_ctry_3 STRING, c_area_3 STRING, c_local_3 STRING, c_ext_3 STRING, email1 STRING, email2 STRING, lcl_tx_id STRING, nat_tx_id STRING, update_dt DATE",
-  "dailymarket": "cdc_flag STRING, cdc_dsn BIGINT, dm_date DATE, dm_s_symb STRING, dm_close DOUBLE, dm_high DOUBLE, dm_low DOUBLE, dm_vol INT",
-  "holdings": "cdc_flag STRING, cdc_dsn BIGINT, hh_h_t_id BIGINT, hh_t_id BIGINT, hh_before_qty INT, hh_after_qty INT, event_dt DATE",
-  "trade": "cdc_flag STRING, cdc_dsn BIGINT, tradeid BIGINT, t_dts TIMESTAMP, status STRING, t_tt_id STRING, cashflag TINYINT, t_s_symb STRING, quantity INT, bidprice DOUBLE, t_ca_id BIGINT, executedby STRING, tradeprice DOUBLE, fee DOUBLE, commission DOUBLE, tax DOUBLE, event_dt DATE",
-  "watches": "cdc_flag STRING, cdc_dsn BIGINT, w_c_id BIGINT, w_s_symb STRING, w_dts TIMESTAMP, w_action STRING, event_dt DATE"
-}
-
-# simulate_filedrops moves the single per-day part file into the watch
-# dir with renamed target `{Dataset}.txt`. After repartition(_pdate)
-# Spark guarantees one part file per date, so no glob is needed.
-file_names = {
-  "account": "Account.txt",
-  "cashtransaction": "CashTransaction.txt",
-  "customer": "Customer.txt",
-  "dailymarket": "DailyMarket.txt",
-  "holdings": "HoldingHistory.txt",
-  "trade": "Trade.txt",
-  "watches": "WatchHistory.txt"
-}
-
-partitions = {
-  "account": ["update_dt"],
-  "cashtransaction": ["event_dt"],
-  "customer": ["update_dt"],
-  "dailymarket": ["dm_date"],
-  "holdings": ["event_dt"],
-  "trade": ["event_dt"],
-  "watches": ["event_dt"]
-}
-
-# COMMAND ----------
-
-def build_autoloader_stream(tbl):
-  return (
-    spark.readStream.format("cloudFiles")
-      .option("cloudFiles.format", "csv")
-      .option("header", "false")
-      .option("sep", "|")
-      .schema(schemas.get(tbl))
-      .option("pathGlobfilter", file_names.get(tbl))
-      #.option("cloudFiles.maxFilesPerTrigger", 1)
-      .load(batches_dir)
-  )
-
-def generate_tables(tbl):
-  tbl_name = (f"bronze{tbl}")
-  @dlt.table(
-    name=tbl_name,
-    cluster_by=partitions.get(tbl),
-    table_properties={"delta.dataSkippingNumIndexedCols": "34"}
-  )
-  def create_table(): 
-    return build_autoloader_stream(tbl)
-
-# COMMAND ----------
-
-for table in tbl_ls:
-  generate_tables(table)
-
-# COMMAND ----------
-
-# Seed bronzecashtransaction with pre-2015-07-06 events so the running
-# `sum(ct_amt) over (partition by accountid order by ct_date)` in the
-# FactCashBalances incremental computes against the correct historical
-# baseline. SDP only — Cluster path uses `currentaccountbalances`
-# directly. Read straight from the shared staging schema (no CLONE
-# needed: SDP can't incrementally update CLONEd tables, and this is
-# always emitted as a once-MV so the staging-schema read is fine).
-# `cashtransactionhistorical` is created by
-# historical/FactCashBalancesHistorical.sql from the temp Delta
-# `tpcdi_raw_data.cashtransaction{sf}` (cleanup_stage0 drops the temp
-# Delta but the staging table persists).
-catalog = spark.conf.get("catalog")
-@dlt.append_flow(
-  target = "bronzecashtransaction",
-  once = True,
-  name = 'bronzecashtransaction_backfill')
-def backfill():
-  return spark.sql(f"""
-    SELECT * FROM {catalog}.tpcdi_incremental_staging_{scale_factor}.cashtransactionhistorical
-  """)
-
-# Seed bronzedailymarket with the prior year (2015-07-06 → 2016-07-05) of
-# DailyMarket data so factmarkethistorystg's 380-day rolling-year MV is
-# fully populated from the first incremental update. Without this seed,
-# factmarkethistorystg would show empty arrays on batch 1 and grow them
-# day-by-day, meaning factmarkethistory's 52-week high/low would be
-# wrong until ~365 days into the run.
-# `bronzedailymarket` (in the staging schema) is created by
-# historical/DailyMarketHistorical.sql from the temp Delta
-# `tpcdi_raw_data.dailymarket{sf}` (cleanup_stage0 drops the temp Delta
-# but the staging table persists).
-@dlt.append_flow(
-  target = "bronzedailymarket",
-  once = True,
-  name = 'bronzedailymarket_backfill')
-def bronzedailymarket_backfill():
-  return spark.sql(f"""
-    SELECT * FROM {catalog}.tpcdi_incremental_staging_{scale_factor}.bronzedailymarket
-  """)
