+++ src/incremental_batches/augmented_incremental/historical/DimAccountHistorical.sql
@@ -22,7 +22,7 @@ CREATE OR REPLACE TABLE DimAccount (
   CONSTRAINT dimaccount_customer_fk FOREIGN KEY (sk_customerid) REFERENCES DimCustomer(sk_customerid),
   CONSTRAINT dimaccount_broker_fk FOREIGN KEY (sk_brokerid) REFERENCES  DimBroker(sk_brokerid)
 )
-CLUSTER BY (enddate)  -- liquid: matches setup_liquid / setup_dbt_liquid (current rows = enddate=9999-12-31 cluster together)
+CLUSTER BY (enddate)  -- liquid: matches setup / setup_dbt (current rows = enddate=9999-12-31 cluster together)
 TBLPROPERTIES (
   'delta.autoOptimize.autoCompact' = 'true',
   'delta.autoOptimize.optimizeWrite' = 'true'
