+++ src/incremental_batches/augmented_incremental/historical/FactWatchesHistorical.sql
@@ -21,7 +21,7 @@ CREATE OR REPLACE TABLE factwatches (
   CONSTRAINT factwatches_dateplaced_fk FOREIGN KEY (sk_dateid_dateplaced) REFERENCES DimDate(sk_dateid),
   CONSTRAINT factwatches_dateremoved_fk FOREIGN KEY (sk_dateid_dateremoved) REFERENCES DimDate(sk_dateid)
 )
-CLUSTER BY (sk_dateid_dateremoved)  -- liquid: matches setup_liquid choice. NOT the `removed` boolean — the same sk_dateid_dateremoved the SDP/cluster liquid variants cluster on
+CLUSTER BY (sk_dateid_dateremoved)  -- liquid: matches setup choice. NOT the `removed` boolean — the same sk_dateid_dateremoved the SDP/cluster liquid variants cluster on
 TBLPROPERTIES (
   'delta.autoOptimize.autoCompact' = 'true',
   'delta.autoOptimize.optimizeWrite' = 'true'
